Move init to own file
