+++ chain/cmd/bandd/main.go
@@ -2,31 +2,20 @@ package main
 
 import (
 	"encoding/json"
-	"fmt"
 	"io"
-	"os"
-	"path/filepath"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/server"
 	"github.com/cosmos/cosmos-sdk/store"
 	"github.com/cosmos/cosmos-sdk/x/genaccounts"
 	genaccscli "github.com/cosmos/cosmos-sdk/x/genaccounts/client/cli"
-	"github.com/cosmos/cosmos-sdk/x/genutil"
 	"github.com/cosmos/cosmos-sdk/x/staking"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	cfg "github.com/tendermint/tendermint/config"
-	"github.com/tendermint/tendermint/crypto/secp256k1"
 	"github.com/tendermint/tendermint/libs/cli"
-	"github.com/tendermint/tendermint/libs/common"
-	cmn "github.com/tendermint/tendermint/libs/common"
 	"github.com/tendermint/tendermint/libs/log"
-	"github.com/tendermint/tendermint/privval"
-	"github.com/tendermint/tendermint/types"
 
 	"github.com/bandprotocol/d3n/chain/app"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -36,155 +25,9 @@ import (
 	dbm "github.com/tendermint/tm-db"
 )
 
-func GenFilePVIfNotExists(cdc *codec.Codec, keyFilePath, stateFilePath string) {
-	if !cmn.FileExists(keyFilePath) {
-		privKey := secp256k1.GenPrivKey()
-		pv := &privval.FilePV{
-			Key: privval.FilePVKey{
-				Address: privKey.PubKey().Address(),
-				PubKey:  privKey.PubKey(),
-				PrivKey: privKey,
-			},
-			LastSignState: privval.FilePVLastSignState{
-				Step: 0,
-			},
-		}
-
-		jsonBytes, err := cdc.MarshalJSONIndent(pv.Key, "", "  ")
-		if err != nil {
-			panic(err)
-		}
-		err = cmn.WriteFileAtomic(keyFilePath, jsonBytes, 0600)
-		if err != nil {
-			panic(err)
-		}
-
-		jsonBytes, err = cdc.MarshalJSONIndent(pv.LastSignState, "", "  ")
-		if err != nil {
-			panic(err)
-		}
-		err = cmn.WriteFileAtomic(stateFilePath, jsonBytes, 0600)
-		if err != nil {
-			panic(err)
-		}
-	}
-}
+const flagInvCheckPeriod = "inv-check-period"
 
 var invCheckPeriod uint
-
-const (
-	flagInvCheckPeriod = "inv-check-period"
-	flagOverwrite      = "overwrite"
-	flagClientHome     = "home-client"
-)
-
-type printInfo struct {
-	Moniker    string          `json:"moniker" yaml:"moniker"`
-	ChainID    string          `json:"chain_id" yaml:"chain_id"`
-	NodeID     string          `json:"node_id" yaml:"node_id"`
-	GenTxsDir  string          `json:"gentxs_dir" yaml:"gentxs_dir"`
-	AppMessage json.RawMessage `json:"app_message" yaml:"app_message"`
-}
-
-func newPrintInfo(moniker, chainID, nodeID, genTxsDir string,
-	appMessage json.RawMessage) printInfo {
-
-	return printInfo{
-		Moniker:    moniker,
-		ChainID:    chainID,
-		NodeID:     nodeID,
-		GenTxsDir:  genTxsDir,
-		AppMessage: appMessage,
-	}
-}
-
-func displayInfo(cdc *codec.Codec, info printInfo) error {
-	out, err := codec.MarshalJSONIndent(cdc, info)
-	if err != nil {
-		return err
-	}
-
-	_, err = fmt.Fprintf(os.Stderr, "%s\n", string(sdk.MustSortJSON(out)))
-	return err
-}
-
-// InitCmd returns a command that initializes all files needed for Tendermint
-// and the respective application.
-func InitCmd(ctx *server.Context, cdc *codec.Codec, customAppState map[string]json.RawMessage,
-	defaultNodeHome string) *cobra.Command { // nolint: golint
-	cmd := &cobra.Command{
-		Use:   "init [moniker]",
-		Short: "Initialize private validator, p2p, genesis, and application configuration files",
-		Long:  `Initialize validators's and node's configuration files.`,
-		Args:  cobra.ExactArgs(1),
-		RunE: func(_ *cobra.Command, args []string) error {
-			config := ctx.Config
-			config.SetRoot(viper.GetString(cli.HomeFlag))
-
-			chainID := viper.GetString(client.FlagChainID)
-			if chainID == "" {
-				chainID = fmt.Sprintf("test-chain-%v", common.RandStr(6))
-			}
-
-			nodeID, _, err := genutil.InitializeNodeValidatorFiles(config)
-			if err != nil {
-				return err
-			}
-
-			config.Moniker = args[0]
-
-			genFile := config.GenesisFile()
-			if !viper.GetBool(flagOverwrite) && common.FileExists(genFile) {
-				return fmt.Errorf("genesis.json file already exists: %v", genFile)
-			}
-			appState, err := codec.MarshalJSONIndent(cdc, customAppState)
-			if err != nil {
-				return err
-			}
-
-			genDoc := &types.GenesisDoc{}
-			if _, err := os.Stat(genFile); err != nil {
-				if !os.IsNotExist(err) {
-					return err
-				}
-			} else {
-				genDoc, err = types.GenesisDocFromFile(genFile)
-				if err != nil {
-					return err
-				}
-			}
-
-			genDoc.ChainID = chainID
-			genDoc.Validators = nil
-			genDoc.AppState = appState
-			genDoc.ConsensusParams = &types.ConsensusParams{
-				Block: types.BlockParams{
-					MaxBytes:   200000, // 0.2MB
-					MaxGas:     2000000,
-					TimeIotaMs: 1000, // 1s
-				},
-				Evidence: types.EvidenceParams{
-					MaxAge: 100000, // 27.8 hrs at 1block/s
-				},
-				Validator: types.ValidatorParams{PubKeyTypes: []string{types.ABCIPubKeyTypeSecp256k1}},
-			}
-			if err = genutil.ExportGenesisFile(genDoc, genFile); err != nil {
-				return err
-			}
-
-			toPrint := newPrintInfo(config.Moniker, chainID, nodeID, "", appState)
-
-			cfg.WriteConfigFile(filepath.Join(config.RootDir, "config", "config.toml"), config)
-			return displayInfo(cdc, toPrint)
-		},
-	}
-
-	cmd.Flags().String(cli.HomeFlag, defaultNodeHome, "node's home directory")
-	cmd.Flags().BoolP(flagOverwrite, "o", false, "overwrite the genesis.json file")
-	cmd.Flags().String(client.FlagChainID, "", "genesis file chain-id, if left blank will be randomly created")
-
-	return cmd
-}
 
 func main() {
 	cobra.EnableCommandSorting = false
